+++ swift/PiqleyPluginSDK/Errors.swift
@@ -5,5 +5,6 @@ public enum SDKError: Error, Sendable {
     case stdinReadFailed
     case payloadDecodeFailed(String)
     case unknownHook(String)
+    case unhandledHook(String)
     case manifestValidationFailed([String])
 }